Extracted logic into a MongoDBModifier
