+++ src/Dashboard/Dashboard.csproj
@@ -156,7 +156,7 @@
     <Compile Include="Infrastructure\MongoDbProvider.cs" />
     <Compile Include="ReadModel\Providers\RAGWidgetViewProvider.cs" />
     <Compile Include="ReadModel\Views\DashboardView.cs" />
-    <Compile Include="Test.cs" />
+    <Compile Include="MongoDbModifier.cs" />
   </ItemGroup>
   <ItemGroup />
   <PropertyGroup>
